[Update] 목표, 장소 삭제 시 관련 미션도 사라지도록 수정

+++ components/Categories.js
@@ -7,6 +7,7 @@ import {
   Text,
   TextInput,
   Modal,
+  Alert,
 } from 'react-native';
 import {useDispatch, useSelector} from 'react-redux';
 import {
@@ -17,7 +18,7 @@ import {
 } from '../store/action';
 import Colors from '../utils/Colors';
 import {useAuth} from '../providers/AuthProvider';
-import {Goal} from '../schema';
+import {Goal, Mission, Place} from '../schema';
 import {createGoalInRealm, deleteGoalInRealm} from '../functions';
 import {mkConfig} from '../functions/mkConfig';
 import Realm from 'realm';
@@ -109,19 +110,40 @@ export default function Categories() {
                 dispatch(selectCategory(item.name));
               }}
               onLongPress={() => {
-                deleteGoalInRealm(user, item.name);
+                Alert.alert(
+                  '목표 삭제',
+                  `목표 "${item.name}" 을/를 삭제하시겠습니까?\n연관된 미션 또한 함께 삭제됩니다. `,
+                  [
+                    {
+                      text: '삭제',
+                      onPress: () => {
+                        Realm.open(
+                          mkConfig(user, [
+                            Goal.schema,
+                            Mission.schema,
+                            Place.schema,
+                          ]),
+                        ).then(async realm => {
+                          await deleteGoalInRealm(user, realm, item);
+                          realm.close();
+                        });
 
-                dispatch(
-                  deleteMission(
-                    mission.filter(el => el.category !== item.name),
-                  ),
+                        dispatch(
+                          deleteMission(
+                            mission.filter(el => el.category !== item.name),
+                          ),
+                        );
+                        dispatch(
+                          deleteCategory(
+                            data.filter(el => el._id !== item._id),
+                          ),
+                        );
+                        dispatch(selectCategory('⭐ 전체 목표'));
+                      },
+                    },
+                    {text: '취소'},
+                  ],
                 );
-                dispatch(
-                  deleteCategory(data.filter(el => el._id !== item._id)),
-                );
-                dispatch(selectCategory('⭐ 전체 목표'));
-
-                // 삭제된 카테고리 관련 미션도 삭제
               }}>
               <Text
                 style={{
